Add support for both direction dependent and independent sizes

+++ src/main/java/it/polimi/ingsw/view/cli/fancy_cli/utils/ConsoleCli.java
@@ -18,7 +18,7 @@ public class ConsoleCli {
      * @param width the width in general points
      * @return the number of characters required
      */
-    static public int convertFromGeneralWidthToCharNumber(int width){
+    static public int convertFromGeneralWidthToCharNumber(float width){
         return Math.round(width*3.3f);
     }
